proxy_spider/requirements.txt
aniso8601==1.2.0
APScheduler==3.3.1
attrs==20.2.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
cycler==0.10.0
fake-useragent==0.1.11
Flask==0.12.1
Flask-BasicAuth==0.2.0
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
Flask-SQLAlchemy==2.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.1
itemloaders==1.0.3
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.10.0
kiwisolver==1.2.0
lxml==4.5.2
MarkupSafe==1.0
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.4.0.44
parsel==1.6.0
Pillow==7.2.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.7.11
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.6.0
pytz==2017.2
queuelib==1.5.0
redis==3.5.3
requests==2.13.0
Scrapy==2.3.0
scrapy-splash==0.7.2
scrapyd==1.2.1
service-identity==18.1.0
six==1.10.0
SpiderKeeper==1.2.0
SQLAlchemy==1.1.9
Twisted==20.3.0
tzlocal==1.3
urllib3==1.25.10
w3lib==1.22.0
Werkzeug==0.12.1
wincertstore==0.2
zope.interface==5.1.1
